IEEE Paper
This paper explores the dual objectives of a
Python-based Hangman game, addressing challenges in digital
game design and language skill enhancement. The Hangman
game aims to preserve the essence of the traditional game while
ensuring an engaging and user-friendly digital rendition.
Furthermore, it endeavours to serve as a practical tool for
bolstering language skills, enriching vocabulary, and fostering
cognitive abilities.